ingestion/batch/update_users/pyproject.toml
[project]
name = "financial-transactions-analytics-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.0",
]
